Change type of one field from int to long and change another files where it needed

+++ VictoryCenter/VictoryCenter.BLL/DTOs/Admin/ReportMediaSettings/CollectedFundsBlockDto.cs
@@ -5,6 +5,6 @@ namespace VictoryCenter.BLL.DTOs.Admin.ReportMediaSettings;
 public record CollectedFundsBlockDto
 {
     public string Title { get; init; } = null!;
-    public int CollectedAmount { get; init; }
+    public long CollectedAmount { get; init; }
     public ImageDto? Image { get; init; }
 }
