+++ VictoryCenter/VictoryCenter.BLL/DTOs/Admin/ReportMediaSettings/UpdateCollectedFundsBlockDto.cs
@@ -3,6 +3,6 @@ namespace VictoryCenter.BLL.DTOs.Admin.ReportMediaSettings;
 public record UpdateCollectedFundsBlockDto
 {
     public string Title { get; init; } = null!;
-    public int CollectedFunds { get; init; }
+    public long CollectedFunds { get; init; }
     public long ImageId { get; init; }
 }
