+++ VictoryCenter/VictoryCenter.DAL/Migrations/VictoryCenterDbContextModelSnapshot.cs
@@ -284,8 +284,8 @@ namespace VictoryCenter.DAL.Migrations
 
                     SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"));
 
-                    b.Property<int>("CollectedAmount")
-                        .HasColumnType("int");
+                    b.Property<long>("CollectedAmount")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTimeOffset>("CreatedAt")
                         .HasColumnType("datetimeoffset");
